refactored alerts so tha each feature has its own file with logic to render alert component
